+++ src/settings/Options.js
@@ -21,6 +21,8 @@ const DEFAULT_OPTIONS = {
   host: "0.0.0.0",
   port: 3100,
   log: "info",
+  cors: true,
+  corsPreFlight: true,
   // TODO, remove v1 legacy code
   behavior: null,
 };